remove usage of <Image /> component
removed Astro <Image /> and replaced with HTML <img>
as (1) remote images wouldn't be optimized
and (2) the element extraction from HTML into component makes things
complicated.

as a result, the <Content /> component is replaced
with a single <article />, and the extraction logic
is removed.

but at some point the images should still be processed
through an Astro integration or something else.